+++ projects/ids-enterprise-ng/src/lib/datagrid/soho-datagrid.component.ts
@@ -213,6 +213,37 @@ export class SohoDataGridComponent implements OnInit, AfterViewInit, OnDestroy, 
   }
 
   /**
+   * The value of the frozenColumns option - returns the requested
+   * value if the control has not been created yet.
+   */
+  get frozenColumns(): SohoDataGridFrozenColumns {
+    if (this.datagrid) {
+      return this.datagrid.settings.frozenColumns;
+    }
+
+    // ... we've been called before the component has completed
+    // initialisation, so return the current value from the
+    // options.
+    return this._gridOptions.frozenColumns;
+  }
+
+  /**
+   * Sets the frozenColumns settings - will force a grid rebuild if the component has already been
+   * created.
+   *
+   * @param frozenColumns - the frozenColumns settings.
+   */
+  @Input() set frozenColumns(frozenColumns: SohoDataGridFrozenColumns) {
+    this._gridOptions.frozenColumns = frozenColumns;
+    if (this.datagrid) {
+      this.datagrid.settings.frozenColumns = frozenColumns;
+
+      // Force all a full rebuild of the control.
+      this.markForRefresh('frozenColumns', RefreshHintFlags.Rebuild);
+    }
+  }
+
+  /**
    * If true allows
    */
   @Input() set cellNavigation(cellNavigation: boolean) {
@@ -1028,12 +1059,27 @@ export class SohoDataGridComponent implements OnInit, AfterViewInit, OnDestroy, 
   @Output()
   expandrow = new EventEmitter<SohoDataGridToggleRowEvent>();
 
+  // This event is fired when edit mode is exited.
+  @Output()
+  exiteditmode = new EventEmitter<SohoDataGridEditModeEvent>();
+
+  // This event is fired before edit mode is started.
+  @Output()
+  beforeentereditmode = new EventEmitter<SohoDataGridEditModeEvent>();
+
+  // This event is fired when edit mode is entered.
+  @Output()
+  entereditmode = new EventEmitter<SohoDataGridEditModeEvent>();
+
   // This event is fired when a row in the grid is collapsed.
   @Output()
   collapserow = new EventEmitter<SohoDataGridToggleRowEvent>();
 
   @Output()
   sorted = new EventEmitter<SohoDataGridSortedEvent>();
+
+  @Output()
+  beforeRowActivated = new EventEmitter<SohoDataGridRowActivated>();
 
   @Output()
   rowActivated = new EventEmitter<SohoDataGridRowActivated>();
@@ -1419,7 +1465,9 @@ export class SohoDataGridComponent implements OnInit, AfterViewInit, OnDestroy, 
    */
   clearFilter(): void {
     this.ngZone.runOutsideAngular(() => {
-      this.datagrid.clearFilter();
+      if (this.datagrid) {
+        this.datagrid.clearFilter();
+      }
     });
   }
 
@@ -1628,6 +1676,24 @@ export class SohoDataGridComponent implements OnInit, AfterViewInit, OnDestroy, 
   }
 
   /**
+   * Reset columns to their defaults (used on restore menu item).
+   */
+  resetColumns(): void {
+    return this.ngZone.runOutsideAngular(() => {
+      this.datagrid.resetColumns();
+    });
+  }
+
+  /**
+   * Open the personalize dialog.
+   */
+  personalizeColumns(): void {
+    return this.ngZone.runOutsideAngular(() => {
+      this.datagrid.personalizeColumns();
+    });
+  }
+
+  /**
    * Restore the user settings from local Storage or as passed in.
    * @param settings The object containing the settings to use.
    */
@@ -1665,6 +1731,39 @@ export class SohoDataGridComponent implements OnInit, AfterViewInit, OnDestroy, 
     const event = { grid: this, ...args };
     this.ngZone.run(() => {
       this.expandrow.next(event);
+    });
+  }
+
+  /**
+   * Event fired after edit mode is activated on an editor.
+   * @param args the event arguments
+   */
+  private onExitEditMode(args: SohoDataGridEditModeEvent) {
+    const event = { grid: this, ...args };
+    this.ngZone.run(() => {
+      this.exiteditmode.next(event);
+    });
+  }
+
+  /**
+   * Event fired before edit mode is activated on an editor.
+   * @param args the event arguments
+   */
+  private onBeforeEnterEditMode(args: SohoDataGridEditModeEvent) {
+    const event = { grid: this, ...args };
+    this.ngZone.run(() => {
+      this.beforeentereditmode.next(event);
+    });
+  }
+
+  /**
+   * Event fired when edit mode is activated on an editor.
+   * @param args the event arguments
+   */
+  private onEnterEditMode(args: SohoDataGridEditModeEvent) {
+    const event = { grid: this, ...args };
+    this.ngZone.run(() => {
+      this.entereditmode.next(event);
     });
   }
 
@@ -1771,9 +1870,18 @@ export class SohoDataGridComponent implements OnInit, AfterViewInit, OnDestroy, 
   /**
    * Event fired when the data is filtered.
    */
-  private onAfterRendere(args: SohoDataGridAfterRenderEvent) {
+  private onAfterRender(args: SohoDataGridAfterRenderEvent) {
     this.ngZone.run(() => {
       this.afterRender.next(args);
+    });
+  }
+
+  /**
+   * Event fired before a row is activated.
+   */
+  private onBeforeRowActivated(args: SohoDataGridRowActivatedEvent) {
+    this.ngZone.run(() => {
+      this.beforeRowActivated.next(args);
     });
   }
 
@@ -1828,6 +1936,31 @@ export class SohoDataGridComponent implements OnInit, AfterViewInit, OnDestroy, 
   private onSorted(args: SohoDataGridSortedEvent) {
     this.ngZone.run(() => {
       this.sorted.next(args);
+    });
+  }
+
+  /**
+   * Returns the row dom jQuery node.
+   * @param  row The row index.
+   * @param  includeGroups If true groups are taken into account.
+   * @return The dom jQuery node
+   */
+  rowNode(row: number, includeGroups: boolean): any {
+    return this.ngZone.runOutsideAngular(() => {
+      return this.datagrid.rowNode(row, includeGroups);
+    });
+  }
+
+  /**
+   * Returns the cell dom node.
+   * @param  row The row index.
+   * @param  cell The cell index.
+   * @param  includeGroups If true groups are taken into account.
+   * @return The dom node
+   */
+  cellNode(row: number, cell: number, includeGroups: boolean): any {
+    return this.ngZone.runOutsideAngular(() => {
+      return this.datagrid.cellNode(row, cell, includeGroups);
     });
   }
 
@@ -2018,12 +2151,16 @@ export class SohoDataGridComponent implements OnInit, AfterViewInit, OnDestroy, 
         .on('collapserow', (e: any, args: SohoDataGridRowCollapseEvent) => { this.onCollapseRow(args); })
         .on('contextmenu', (e: any, args: SohoDataGridRowClicked) => { this.onContextMenu(args); })
         .on('dblclick', (e: JQuery.TriggeredEvent, args: SohoDataGridRowClicked) => { this.onDoubleClick(args); })
+        .on('beforeentereditmode', (e: any, args: SohoDataGridEditModeEvent) => { this.onBeforeEnterEditMode(args); })
+        .on('exiteditmode', (e: any, args: SohoDataGridEditModeEvent) => { this.onExitEditMode(args); })
+        .on('entereditmode', (e: any, args: SohoDataGridEditModeEvent) => { this.onEnterEditMode(args); })
         .on('expandrow', (e: any, args: SohoDataGridRowExpandEvent) => { this.onExpandRow(args); })
         .on('filtered', (e: any, args: SohoDataGridFilteredEvent) => { this.onFiltered(args); })
         .on('openfilterrow', (e: any, args: SohoDataGridOpenFilterRowEvent) => { this.onOpenFilterRow(args); })
         .on('rowremove', (e: any, args: SohoDataGridRowRemoveEvent) => { this.onRowRemove(args); })
         .on('rendered', (e: any, args: SohoDataGridRenderedEvent) => { this.onRendered(args); })
-        .on('afterrender', (e: any, args: SohoDataGridAfterRenderEvent) => { this.onAfterRendere(args); })
+        .on('afterrender', (e: any, args: SohoDataGridAfterRenderEvent) => { this.onAfterRender(args); })
+        .on('beforerowactivated', (e: any, args: SohoDataGridRowActivatedEvent) => { this.onBeforeRowActivated(args); })
         .on('rowactivated', (e: any, args: SohoDataGridRowActivatedEvent) => { this.onRowActivated(args); })
         .on('rowdeactivated', (e: any, args: SohoDataGridRowDeactivatedEvent) => { this.onRowDeactivated(args); })
         .on('rowreorder', (e: any, args: SohoDataGridRowReorderedEvent) => { this.onRowReordered(args); })
@@ -2143,24 +2280,6 @@ export class SohoDataGridComponent implements OnInit, AfterViewInit, OnDestroy, 
       }
     });
   }
-}
-
-/**
- * Enumeration of the Soho Datagrid FilterTypes.
- * Note: integer, percent and lookup are not yet implemented the soho datagrid.js
- * See http://stackoverflow.com/questions/15490560/create-an-enum-with-string-values-in-typescript
- * for more details about creating an enum of strings.
- */
-export enum SohoGridColumnFilterTypes {
-  Text     = 'text',
-  Checkbox = 'checkbox',
-  Contents = 'contents',
-  Date     = 'date',
-  Decimal  = 'decimal',
-  Integer  = 'integer',
-  Lookup   = 'lookup',
-  Percent  = 'percent',
-  Select   = 'select'
 }
 
 /**