+++ src/app/personalize-menu/theme.model.ts
@@ -1,0 +1,5 @@
+export class Theme {
+  name: string;
+  value: string;
+  selected?: boolean;
+}